refactor(owner-view): owner detail = public layout + manage bar (products, slice 1)
The owner dashboard detail page was a flat label→value grid (EntityDetailPage),
teaching owners their listing looks like a database row. Per the entity-detail
redesign, the owner should see exactly what buyers see, plus management controls.

- PublicEntityDetailPage is now owner-aware: detects the viewer owns the entity
  (published OR draft) and renders a manage bar (status note + Edit). The earlier
  draft-only preview banner folds into it. Buyers see no bar.
- Extracted productDetailConfig to a shared module (SSOT) consumed by both the
  public route (/products/[id]) and the owner route (/dashboard/store/[id]).
- /dashboard/store/[id] now renders PublicEntityDetailPage — same layout as the
  public page, with the owner manage bar — replacing the flat grid + makeDetailFields.

Slice 1 of N: products. Services/projects/causes/research/ai_assistants (already on
PublicEntityDetailPage) can follow the same pattern; the 6 entity types still on
EntityDetailPage for their PUBLIC pages (loans/investments/assets/wishlists/circles/
groups) need their own PublicEntityDetailPage configs — tracked as follow-ups.

tsc 0; lint clean.
